changes for sharepoint integration
