+++ src/api/common/constants/queue-initial-data.js
@@ -1,0 +1,40 @@
+export const messages = [
+  {
+    fileName: 'SBI_FRN_CPH.xlsx',
+    filePath: '/Selection/FETF/SBI_FRN_CPH.xlsx'
+  },
+  {
+    fileName: 'CPH_SEO_Group_Look_Up_Table_V4_23.01.2025.xlsx',
+    filePath: '/Selection/FETF/CPH_SEO_Group_Look_Up_Table_V4_23.01.2025.xlsx'
+  },
+  {
+    fileName: 'giles_report_official_sensitive_1.xlsb',
+    filePath: '/Selection/FETF/giles_report_official_sensitive_1.xlsb'
+  },
+  {
+    fileName: 'giles_report_official_sensitive_2b.xlsb',
+    filePath: '/Selection/FETF/giles_report_official_sensitive_2b.xls.xlsb'
+  },
+  {
+    fileName: 'CS_MEASURES.xlsb',
+    filePath: '/Selection/FETF/CS_MEASURES.xlsb'
+  }
+]
+
+export const queueInitialInfo = [
+  {
+    fileName: 'SBI_FRN_CPH.xlsx'
+  },
+  {
+    fileName: 'CPH_SEO_Group_Look_Up_Table_V4_23.01.2025.xlsx'
+  },
+  {
+    fileName: 'giles_report_official_sensitive_1.xlsb'
+  },
+  {
+    fileName: 'giles_report_official_sensitive_2b.xlsb'
+  },
+  {
+    fileName: 'CS_MEASURES.xlsb'
+  }
+]
